requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
docopt==0.6.2
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Genshi==0.7.7
google-search-results==2.4.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonpickle==3.0.1
MarkupSafe==2.1.2
numpy==1.24.2
pandas==2.0.0
pipreqs==0.4.11
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.7
threadpoolctl==3.1.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.2
Whoosh==2.7.4
yarg==0.1.9
